[vm] Add a register-preserving calling convention
The PCALL and PRET args utilize the base pointer register (RBP) to provide a higher-level (albeit more expensive) calling convention.
While the details of operation can be read up in the code, a simple description is as follows:
-  at call time, PCALL preserves all registers as if the caller had PUSHed them individually before PUSHing args and CALLing.
-  at ret time, PRET restores all registers, leaving return values in the current frame , ie - in [RBP, RSP).

The stack management done by PCALL and PRET is transparent to programmers.